Facenet/requirements.txt
numpy==1.16.4
Pillow==9.0.1
matplotlib==3.1.0
opencv-contrib-python==4.2.0.32
scipy==1.2.1
scikit-learn==0.21.2
tensorflow==1.14.0
Keras==2.2.4